+++ api/api.go
@@ -1,4 +1,22 @@
 /*
+ *  Nuts consent logic holds the logic for consent creation
+ *  Copyright (C) 2019 Nuts community
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+/*
  * This file is part of nuts-consent-logic.
  *
  * nuts-consent-logic is free software: you can redistribute it and/or modify
@@ -19,10 +37,11 @@
 package api
 
 import (
-	"github.com/labstack/echo/v4"
-	"github.com/nuts-foundation/nuts-consent-logic/pkg"
 	"net/http"
 	"time"
+
+	"github.com/labstack/echo/v4"
+	"github.com/nuts-foundation/nuts-consent-logic/pkg"
 )
 
 // Wrapper provides the implementation of the generated ServerInterface
