+++ engine/engine.go
@@ -1,4 +1,22 @@
 /*
+ *  Nuts consent logic holds the logic for consent creation
+ *  Copyright (C) 2019 Nuts community
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+/*
  * This file is part of nuts-consent-logic.
  *
  * nuts-consent-logic is free software: you can redistribute it and/or modify
@@ -19,13 +37,14 @@
 package engine
 
 import (
+	"go/types"
+
 	"github.com/deepmap/oapi-codegen/pkg/runtime"
 	"github.com/nuts-foundation/nuts-consent-logic/api"
 	"github.com/nuts-foundation/nuts-consent-logic/pkg"
 	engine "github.com/nuts-foundation/nuts-go-core"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
-	"go/types"
 )
 
 func NewConsentLogicEngine() *engine.Engine {
