+++ pkg/logic.go
@@ -1,4 +1,22 @@
 /*
+ *  Nuts consent logic holds the logic for consent creation
+ *  Copyright (C) 2019 Nuts community
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+/*
  * This file is part of nuts-consent-logic.
  *
  * nuts-consent-logic is free software: you can redistribute it and/or modify
@@ -26,6 +44,9 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"reflect"
+	"sync"
+
 	bridgeClient "github.com/nuts-foundation/consent-bridge-go-client/api"
 	cStoreClient "github.com/nuts-foundation/nuts-consent-store/client"
 	cStore "github.com/nuts-foundation/nuts-consent-store/pkg"
@@ -39,8 +60,6 @@ import (
 	uuid "github.com/satori/go.uuid"
 	"github.com/sirupsen/logrus"
 	"gopkg.in/thedevsaddam/gojsonq.v2"
-	"reflect"
-	"sync"
 )
 
 type ConsentLogicConfig struct {
@@ -108,13 +127,13 @@ func (cl ConsentLogic) buildConsentRequestConstructedEvent(createConsentRequest 
 	}
 
 	{
-		if res, err := CustodianIsKnown(cl.NutsCrypto, *createConsentRequest); !res || err != nil {
+		if res, err := cl.custodianIsKnown(*createConsentRequest); !res || err != nil {
 			return nil, errors.New("custodian is not a known vendor")
 		}
 		logger().Debug("Custodian is known")
 	}
 	{
-		if consentID, err = GetConsentId(cl.NutsCrypto, *createConsentRequest); consentID == "" || err != nil {
+		if consentID, err = cl.getConsentID(*createConsentRequest); consentID == "" || err != nil {
 			fmt.Println(err)
 			// todo: report back the reason why the consentID could not be generated. Probably because the custodian is not managed by this node?
 			return nil, errors.New("could not create the consentID for this combination of subject, actor and custodian")
@@ -122,7 +141,7 @@ func (cl ConsentLogic) buildConsentRequestConstructedEvent(createConsentRequest 
 		logger().Debug("ConsentId generated")
 	}
 	{
-		if versionID, err = getVersionID(cl.NutsCrypto, *createConsentRequest); consentID == "" || err != nil {
+		if versionID, err = cl.getVersionID(*createConsentRequest); versionID == "" || err != nil {
 			err = fmt.Errorf("could not determine versionId: %w", err)
 			logger().Error(err)
 			return nil, err
@@ -137,19 +156,19 @@ func (cl ConsentLogic) buildConsentRequestConstructedEvent(createConsentRequest 
 			if createConsentRequest.Performer != nil {
 				performer = *createConsentRequest.Performer
 			}
-			if fhirConsent, err = CreateFhirConsentResource(createConsentRequest.Custodian, createConsentRequest.Actor, createConsentRequest.Subject, performer, record); fhirConsent == "" || err != nil {
+			if fhirConsent, err = cl.createFhirConsentResource(createConsentRequest.Custodian, createConsentRequest.Actor, createConsentRequest.Subject, performer, record); fhirConsent == "" || err != nil {
 				return nil, fmt.Errorf("could not create the FHIR consent resource: %w", err)
 			}
 			logger().Debug("FHIR resource created", fhirConsent)
 		}
 		{
-			if validationResult, err := ValidateFhirConsentResource(fhirConsent); !validationResult || err != nil {
+			if validationResult, err := cl.validateFhirConsentResource(fhirConsent); !validationResult || err != nil {
 				return nil, fmt.Errorf("the generated FHIR consent resource is invalid: %w", err)
 			}
 			logger().Debug("FHIR resource is valid")
 		}
 		{
-			if encryptedConsent, err = EncryptFhirConsent(cl.NutsRegistry, cl.NutsCrypto, fhirConsent, *createConsentRequest); err != nil {
+			if encryptedConsent, err = cl.encryptFhirConsent(fhirConsent, *createConsentRequest); err != nil {
 				return nil, fmt.Errorf("could not encrypt consent resource for all involved parties: %w", err)
 			}
 			logger().Debug("FHIR resource encrypted")
@@ -328,7 +347,7 @@ func (cl ConsentLogic) HandleIncomingCordaEvent(event *events.Event) {
 
 		// validate consent record
 		// =======================
-		if validationResult, err := ValidateFhirConsentResource(fhirConsent); !validationResult || err != nil {
+		if validationResult, err := cl.validateFhirConsentResource(fhirConsent); !validationResult || err != nil {
 			errorDescription := "Consent record invalid"
 			event.Error = &errorDescription
 			logger().WithError(err).Error(errorDescription)
@@ -387,7 +406,7 @@ func (cl ConsentLogic) signConsentRequest(event events.Event) (*events.Event, er
 		consentRecordHash := *cr.AttachmentHash
 		logger().Debugf("signing for LegalEntity %s and consentRecordHash %s", legalEntityToSignFor, consentRecordHash)
 
-		pubKey, err := cl.NutsCrypto.PublicKey(cryptoTypes.LegalEntity{URI: legalEntityToSignFor})
+		pubKey, err := cl.NutsCrypto.PublicKeyInPEM(cryptoTypes.LegalEntity{URI: legalEntityToSignFor})
 		if err != nil {
 			logger().Errorf("Error in getting pubKey for %s: %v", legalEntityToSignFor, err)
 			return nil, err
@@ -491,7 +510,7 @@ func (cl ConsentLogic) findFirstEntityToSignFor(signatures *[]bridgeClient.Party
 		// ... check if it has already signed the request
 		if !attSignatures[string(ent)] {
 			// if not, if this node manages the public key, than it is our identity.
-			key, _ := cl.NutsCrypto.PublicKey(cryptoTypes.LegalEntity{URI: string(ent)})
+			key, _ := cl.NutsCrypto.PublicKeyInPEM(cryptoTypes.LegalEntity{URI: string(ent)})
 			if len(key) != 0 {
 				// yes, so lets add it to the missingSignatures so we can sign it in the next step
 				logger().Debugf("found first entity to sign for: %v", ent)
@@ -519,7 +538,7 @@ func (cl ConsentLogic) autoAckConsentRequest(event events.Event) (*events.Event,
 type FHIRResourceWithHash struct {
 	FHIRResource string
 	// Hash represents the attachment hash (zip of cipherText and metadata) from the distributed event model
-	Hash         string
+	Hash string
 }
 
 // HandleEventConsentDistributed handles EventConsentDistributed.
@@ -541,7 +560,7 @@ func (cl ConsentLogic) HandleEventConsentDistributed(event *events.Event) {
 
 	for _, cr := range crs.ConsentRecords {
 		for _, organisation := range cr.Metadata.OrganisationSecureKeys {
-			publicKey, err := cl.NutsCrypto.PublicKey(cryptoTypes.LegalEntity{URI: string(organisation.LegalEntity)})
+			publicKey, err := cl.NutsCrypto.PublicKeyInPEM(cryptoTypes.LegalEntity{URI: string(organisation.LegalEntity)})
 			if err != nil || publicKey == "" {
 				// this organisation is not managed by this node, try with next
 				continue
@@ -592,11 +611,11 @@ func hashFHIRConsent(fhirConsent string) string {
 // only consent records of which or the custodian or the actor is managed by this node should be stored
 func (cl ConsentLogic) isRelevantForThisNode(patientConsent cStore.PatientConsent) bool {
 	// add if custodian is managed by this node
-	if key, _ := cl.NutsCrypto.PublicKey(cryptoTypes.LegalEntity{URI: patientConsent.Custodian}); key != "" {
+	if key, _ := cl.NutsCrypto.PublicKeyInPEM(cryptoTypes.LegalEntity{URI: patientConsent.Custodian}); key != "" {
 		return true
 	}
 	// or if the actor is managed by this node
-	if key, _ := cl.NutsCrypto.PublicKey(cryptoTypes.LegalEntity{URI: patientConsent.Actor}); key != "" {
+	if key, _ := cl.NutsCrypto.PublicKeyInPEM(cryptoTypes.LegalEntity{URI: patientConsent.Actor}); key != "" {
 		return true
 	}
 	return false
