--- pkg/step-custodian-is-known.go
@@ -1,33 +1,0 @@
-/*
- * This file is part of nuts-consent-logic.
- *
- * nuts-consent-logic is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * nuts-consent-logic is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with nuts-consent-logic.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package pkg
-
-import (
-	"github.com/nuts-foundation/nuts-crypto/pkg"
-	"github.com/nuts-foundation/nuts-crypto/pkg/types"
-)
-
-// CustodianIsKnown checks if the custodian from the request is managed by this node.
-// If not, the request cannot continue.
-func CustodianIsKnown(cryptoClient pkg.Client, request CreateConsentRequest) (bool, error) {
-	if pkey, _ := cryptoClient.PublicKey(types.LegalEntity{URI: string(request.Custodian)}); pkey == "" {
-		return false, nil
-	}
-	return true, nil
-}
